+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.6] / [1.1.10] - 2017-11-14
+### Added
+- Am-Mailing in ApiBootstrap
+- New Password request flow in am-auth
+- md5 and am-mailing Import
+- Fix in find and Update user
+
 ## [1.1.5] - 2017-11-13
 ### Added
 - ApiBootstrap Added